Split utils by using on front and shared
