+++ src/components/Blog/Post/index.tsx
@@ -5,9 +5,9 @@ import { useWindowScroll, useWindowSize } from 'react-use'
 
 import { IBlogPostData } from '../../../templates/blog-post'
 
-import { useCommentsCount } from '../../../utils/api'
-import { pluralizeComments } from '../../../utils/i18n'
-import tagToSlug from '../../../utils/tagToSlug'
+import { useCommentsCount } from '../../../utils/front/api'
+import { pluralizeComments } from '../../../utils/front/i18n'
+import tagToSlug from '../../../utils/shared/tagToSlug'
 
 import Markdown from './Markdown'
 import FeedMeta from '../FeedMeta'
