+++ src/components/Community/Contribute/index.tsx
@@ -5,7 +5,7 @@ import LayoutWidthContainer from '../../LayoutWidthContainer'
 import Link from '../../Link'
 import CommunityBlock from '../Block'
 import CommunitySection from '../Section'
-import { logEvent } from '../../../utils/ga'
+import { logEvent } from '../../../utils/front/ga'
 
 import data from '../data.json'
 import sharedStyles from '../styles.module.css'
