+++ src/components/Community/Learn/index.tsx
@@ -8,11 +8,11 @@ import Link from '../../Link'
 import Block from '../Block'
 import Section from '../Section'
 
-import { logEvent } from '../../../utils/ga'
-import { getFirstPage } from '../../../utils/sidebar'
-import { useCommentsCount } from '../../../utils/api'
+import { logEvent } from '../../../utils/front/ga'
+import { getFirstPage } from '../../../utils/shared/sidebar'
+import { useCommentsCount } from '../../../utils/front/api'
 import getPosts from '../../../queries/posts'
-import { pluralizeComments } from '../../../utils/i18n'
+import { pluralizeComments } from '../../../utils/front/i18n'
 
 import data from '../data.json'
 import sharedStyles from '../styles.module.css'
