+++ src/components/Home/index.tsx
@@ -8,7 +8,7 @@ import LearnMore from './LearnMore'
 import LandingHero from './LandingHero'
 import Diagram from './Diagram'
 import UseCases from './UseCases'
-import { logEvent } from '../../utils/ga'
+import { logEvent } from '../../utils/front/ga'
 
 import styles from './styles.module.css'
