+++ src/components/Page/utils.ts
@@ -1,9 +1,9 @@
 import { useEffect } from 'react'
 import { useLocation } from '@reach/router'
 
-import { handleFrontRedirect } from '../../utils/redirects'
-import { allImagesLoadedInContainer } from '../../utils/images'
-import { scrollIntoLayout } from '../../utils/scroll'
+import { handleFrontRedirect } from '../../utils/shared/redirects'
+import { allImagesLoadedInContainer } from '../../utils/front/images'
+import { scrollIntoLayout } from '../../utils/front/scroll'
 
 import styles from './styles.module.css'
