+++ src/utils/shared/sidebar.test.js
@@ -19,7 +19,7 @@ describe('SidebarMenu/helper', () => {
         }
       ]
 
-      jest.doMock('../../content/docs/sidebar.json', () => rawData)
+      jest.doMock('../../../content/docs/sidebar.json', () => rawData)
       const sidebarData = require('./sidebar').structure
 
       expect(sidebarData).toEqual(result)
@@ -38,7 +38,7 @@ describe('SidebarMenu/helper', () => {
         }
       ]
 
-      jest.doMock('../../content/docs/sidebar.json', () => rawData)
+      jest.doMock('../../../content/docs/sidebar.json', () => rawData)
       const sidebarData = require('./sidebar').structure
 
       expect(sidebarData).toEqual(result)
@@ -57,7 +57,7 @@ describe('SidebarMenu/helper', () => {
         }
       ]
 
-      jest.doMock('../../content/docs/sidebar.json', () => rawData)
+      jest.doMock('../../../content/docs/sidebar.json', () => rawData)
       const sidebarData = require('./sidebar').structure
 
       expect(sidebarData).toEqual(result)
@@ -76,7 +76,7 @@ describe('SidebarMenu/helper', () => {
         }
       ]
 
-      jest.doMock('../../content/docs/sidebar.json', () => rawData)
+      jest.doMock('../../../content/docs/sidebar.json', () => rawData)
       const sidebarData = require('./sidebar').structure
 
       expect(sidebarData).toEqual(result)
@@ -106,7 +106,7 @@ describe('SidebarMenu/helper', () => {
         }
       ]
 
-      jest.doMock('../../content/docs/sidebar.json', () => rawData)
+      jest.doMock('../../../content/docs/sidebar.json', () => rawData)
       const sidebarData = require('./sidebar').structure
 
       expect(sidebarData).toEqual(result)
@@ -150,7 +150,7 @@ describe('SidebarMenu/helper', () => {
         }
       ]
 
-      jest.doMock('../../content/docs/sidebar.json', () => rawData)
+      jest.doMock('../../../content/docs/sidebar.json', () => rawData)
       const sidebarData = require('./sidebar').structure
 
       expect(sidebarData).toEqual(result)
@@ -190,7 +190,7 @@ describe('SidebarMenu/helper', () => {
         }
       ]
 
-      jest.doMock('../../content/docs/sidebar.json', () => rawData)
+      jest.doMock('../../../content/docs/sidebar.json', () => rawData)
       const sidebarData = require('./sidebar').structure
 
       expect(sidebarData).toEqual(result)
@@ -230,7 +230,7 @@ describe('SidebarMenu/helper', () => {
         }
       ]
 
-      jest.doMock('../../content/docs/sidebar.json', () => rawData)
+      jest.doMock('../../../content/docs/sidebar.json', () => rawData)
       const sidebarData = require('./sidebar').structure
 
       expect(sidebarData).toEqual(result)
@@ -286,7 +286,7 @@ describe('SidebarMenu/helper', () => {
         }
       ]
 
-      jest.doMock('../../content/docs/sidebar.json', () => rawData)
+      jest.doMock('../../../content/docs/sidebar.json', () => rawData)
       const sidebarData = require('./sidebar').structure
 
       expect(sidebarData).toEqual(result)
@@ -295,7 +295,7 @@ describe('SidebarMenu/helper', () => {
     it("Throws error if item didn't have slug field", () => {
       const rawData = [{}]
 
-      jest.doMock('../../content/docs/sidebar.json', () => rawData)
+      jest.doMock('../../../content/docs/sidebar.json', () => rawData)
 
       expect(() => require('./sidebar')).toThrow(
         new Error("'slug' field is required in objects in sidebar.json")
@@ -306,7 +306,7 @@ describe('SidebarMenu/helper', () => {
     it("Throws error if item has source: false and doesn't have children", () => {
       const rawData = [{ slug: 'item-name', source: false }]
 
-      jest.doMock('../../content/docs/sidebar.json', () => rawData)
+      jest.doMock('../../../content/docs/sidebar.json', () => rawData)
 
       expect(() => require('./sidebar')).toThrow(
         new Error(
@@ -328,7 +328,7 @@ describe('SidebarMenu/helper', () => {
         next: undefined
       }
 
-      jest.doMock('../../content/docs/sidebar.json', () => rawData)
+      jest.doMock('../../../content/docs/sidebar.json', () => rawData)
       const { getItemByPath } = require('./sidebar')
 
       expect(getItemByPath('/doc')).toEqual(result)
@@ -364,7 +364,7 @@ describe('SidebarMenu/helper', () => {
         next: undefined
       }
 
-      jest.doMock('../../content/docs/sidebar.json', () => rawData)
+      jest.doMock('../../../content/docs/sidebar.json', () => rawData)
       const { getItemByPath } = require('./sidebar')
 
       expect(getItemByPath('/doc/item')).toEqual(result)
@@ -383,7 +383,7 @@ describe('SidebarMenu/helper', () => {
         '/doc/item-name/nested-item/subnested-item'
       ]
 
-      jest.doMock('../../content/docs/sidebar.json', () => rawData)
+      jest.doMock('../../../content/docs/sidebar.json', () => rawData)
       const { getParentsListFromPath } = require('./sidebar')
 
       expect(getParentsListFromPath(path)).toEqual(result)
